Developer update #5 - 24.07.2019
1) Created a separate component (ParticleSettings.js) for Particles.js
2) Created a separate component (Main.js) for the main app functionality
3) App.js component now handles only the composition and the app routing logic
4) Added the basic routing logic
