+++ src/components/imageinputform/imageinputform.css
@@ -7,7 +7,7 @@
     background-color: #FFFFFF;
     color: black;
     font-weight: 700;
-    transition: all ease 1s;
+    transition: all ease 750ms;
     transform: scale(1);
 }
 
@@ -16,6 +16,6 @@
     color: black;
     border: 1px solid yellow;
     font-weight: 700;
-    transition: all ease 1s;
-    transform: scale(1.1)
+    transform: scale(1.1);
+    transition: all ease 200ms;
 }
