+++ src/components/login/login.css
@@ -21,3 +21,7 @@
 a {
     text-decoration: none;
 }
+
+.mw5-5 {
+    max-width: 24rem;
+}
