+++ src/components/particles/particlesettings.css
@@ -1,0 +1,8 @@
+.particles {
+    position: fixed;
+    top: 0;
+    right: 0;
+    bottom: 0;
+    left: 0;
+    z-index: -9999;
+  }
