+++ src/containers/app/App.css
@@ -2,14 +2,7 @@
   text-align: center;
 }
 
-.particles {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  z-index: -9999;
-}
+
